Token class and file were causing issues
With python built-in token class and files
