+++ src/menopy/parser.py
@@ -1,7 +1,6 @@
 from bs4 import BeautifulSoup
 import bs4
 from menopy.constants import MENOTA_ENTITIES_URL
-from menopy.token import Token
 from menopy.menodoc import MeNoDoc
 from menopy.load_menoxml import load_menoxml
 from menopy.head_extractor import extract_head
