--- src/menopy/token.py
@@ -1,19 +1,0 @@
-class Token:
-
-    normalized: str
-    diplomatic: str
-    facsimile: str
-    lemma: str
-    msa: str
-
-    def __init__(self,
-                normalized: str,
-                diplomatic: str,
-                facsimile: str,
-                lemma: str,
-                msa: str) -> None:
-        self.normalized = normalized
-        self.diplomatic = diplomatic
-        self.facsimile = facsimile
-        self.lemma = lemma
-        self.msa = msa
